TRY: move defaultTemplateTypes to core/editor
