+++ packages/edit-site/src/components/navigation-sidebar/navigation-panel/new-template-dropdown.js
@@ -24,7 +24,7 @@ import { TEMPLATES_STATUSES } from './constants';
 
 export default function NewTemplateDropdown() {
 	const { defaultTemplateTypes, templates } = useSelect( ( select ) => {
-		const { getDefaultTemplateTypes } = select( 'core/edit-site' );
+		const { getDefaultTemplateTypes } = select( 'core/editor' );
 		const templateEntities = select( 'core' ).getEntityRecords(
 			'postType',
 			'wp_template',
