+++ packages/edit-site/src/components/template-details/index.js
@@ -12,7 +12,7 @@ import { MENU_TEMPLATES } from '../navigation-sidebar/navigation-panel/constants
 
 export default function TemplateDetails( { template, onClose } ) {
 	const { title, description } = useSelect(
-		( select ) => select( 'core/edit-site' ).getTemplateInfo( template ),
+		( select ) => select( 'core/editor' ).getTemplateInfo( template ),
 		[]
 	);
 	const { openNavigationPanelToMenu } = useDispatch( 'core/edit-site' );
